Rename the checkbox theme overwrites file
