+++ packages/ui/src/theme/overwrites/components/index.ts
@@ -1,5 +1,5 @@
 import { Button } from './button';
-import { Checkbox } from './checkbox-styles';
+import { Checkbox } from './checkbox';
 import { Input } from './input';
 import { NumberInput } from './numberInput';
 import { Radio } from './radio';
